restructures root src/ directory and comments App.js
